+++ prowler/lib/outputs/compliance/cis_aws.py
@@ -1,6 +1,6 @@
 from prowler.config.config import timestamp
+from prowler.lib.outputs.compliance.models import Check_Output_CSV_AWS_CIS
 from prowler.lib.outputs.csv.csv import generate_csv_fields
-from prowler.lib.outputs.models import Check_Output_CSV_AWS_CIS
 from prowler.lib.utils.utils import outputs_unix_timestamp
