+++ prowler/lib/outputs/compliance/iso27001_2013_aws.py
@@ -1,8 +1,8 @@
 from csv import DictWriter
 
 from prowler.config.config import timestamp
+from prowler.lib.outputs.compliance.models import Check_Output_CSV_AWS_ISO27001_2013
 from prowler.lib.outputs.csv.csv import generate_csv_fields
-from prowler.lib.outputs.models import Check_Output_CSV_AWS_ISO27001_2013
 from prowler.lib.utils.utils import outputs_unix_timestamp
