+++ prowler/lib/outputs/compliance/models.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import Optional
 
 from pydantic import BaseModel
 
@@ -179,53 +179,3 @@ class Check_Output_CSV_AWS_ISO27001_2013(BaseModel):
     StatusExtended: str
     ResourceId: str
     CheckId: str
-
-
-# JSON ASFF Output
-class ProductFields(BaseModel):
-    ProviderName: str = "Prowler"
-    ProviderVersion: str
-    ProwlerResourceName: str
-
-
-class Severity(BaseModel):
-    Label: str
-
-
-class Resource(BaseModel):
-    Type: str
-    Id: str
-    Partition: str
-    Region: str
-    Tags: Optional[dict]
-
-
-class Compliance(BaseModel):
-    Status: str
-    RelatedRequirements: List[str]
-    AssociatedStandards: List[dict]
-
-
-# TODO: move this to outputs/json_asff/models.py
-class Check_Output_JSON_ASFF(BaseModel):
-    """
-    Check_Output_JSON_ASFF generates a finding's output in JSON ASFF format: https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-findings-format-syntax.html
-    """
-
-    SchemaVersion: str = "2018-10-08"
-    Id: str = ""
-    ProductArn: str = ""
-    RecordState: str = "ACTIVE"
-    ProductFields: ProductFields = None
-    GeneratorId: str = ""
-    AwsAccountId: str = ""
-    Types: List[str] = None
-    FirstObservedAt: str = ""
-    UpdatedAt: str = ""
-    CreatedAt: str = ""
-    Severity: Severity = None
-    Title: str = ""
-    Description: str = ""
-    Resources: List[Resource] = None
-    Compliance: Compliance = None
-    Remediation: dict = None
