+++ prowler/providers/aws/lib/security_hub/security_hub.py
@@ -4,7 +4,7 @@ from botocore.client import ClientError
 from prowler.config.config import timestamp_utc
 from prowler.lib.logger import logger
 from prowler.lib.outputs.json_asff.json_asff import fill_json_asff
-from prowler.lib.outputs.models import Check_Output_JSON_ASFF
+from prowler.lib.outputs.json_asff.models import Check_Output_JSON_ASFF
 
 SECURITY_HUB_INTEGRATION_NAME = "prowler/prowler"
 SECURITY_HUB_MAX_BATCH = 100
